[feat] apply common api fetcher for product data fetching
